pyproject.toml
[tool.poetry]
name = "pro-1"
version = "0.1.0"
description = ""
authors = ["Abdullah  <abdullahr797@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlmodel = "^0.0.22"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
passlib = "^1.7.4"
bcrypt = "<4.0.0"
pyjwt = "^2.10.1"
starlette = ">=0.40.0,<0.42.0"
psycopg2-binary = "^2.9.10"



[tool.poetry.scripts]
dev = "pro_1.main:start"


[tool.poetry.group.dev.dependencies]
alembic = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
